:hammer: Refactor generateComposite
